Add tests to ArrayChangerParams
