+++ Lesson6DZ/Logging/src/test/java/ArrayOf1And4TestParams.java
@@ -1,0 +1,4 @@
+package PACKAGE_NAME;
+
+public class ArrayOf1And4TestParams {
+}
